// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      ingredients: {
        Row: {
          category: string | null
          id: number
          meal_id: number | null
          name: string
          quantity: string | null
          shopping_link: string | null
        }
        Insert: {
          category?: string | null
          id?: number
          meal_id?: number | null
          name: string
          quantity?: string | null
          shopping_link?: string | null
        }
        Update: {
          category?: string | null
          id?: number
          meal_id?: number | null
          name?: string
          quantity?: string | null
          shopping_link?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "ingredients_meal_id_fkey"
            columns: ["meal_id"]
            isOneToOne: false
            referencedRelation: "meals"
            referencedColumns: ["id"]
          },
        ]
      }
      meals: {
        Row: {
          calories: number | null
          cooking_steps: string[] | null
          description: string | null
          estimated_price: number | null
          estimated_time: string | null
          id: number
          image_hint: string | null
          image_url: string | null
          name: string
          protein: number | null
          video_url: string | null
        }
        Insert: {
          calories?: number | null
          cooking_steps?: string[] | null
          description?: string | null
          estimated_price?: number | null
          estimated_time?: string | null
          id?: number
          image_hint?: string | null
          image_url?: string | null
          name: string
          protein?: number | null
          video_url?: string | null
        }
        Update: {
          calories?: number | null
          cooking_steps?: string[] | null
          description?: string | null
          estimated_price?: number | null
          estimated_time?: string | null
          id?: number
          image_hint?: string | null
          image_url?: string | null
          name?: string
          protein?: number | null
          video_url?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database["public"]["Tables"] & Database["public"]["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database["public"]["Tables"] &
        Database["public"]["Views"])
    ? (Database["public"]["Tables"] &
        Database["public"]["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["public"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["public"]["Enums"]
    ? Database["public"]["Enums"][PublicEnumNameOrOptions]
    : never